:recycle: Extract CliActions from cli.go
